# ~2x faster uncached `bun install` on Windows