--- a/src/install/npm.zig
+++ b/src/install/npm.zig
@@ -708,7 +708,7 @@ pub const PackageManifest = struct {
                 try bun.sys.renameat(bun.FD.cwd(), tmp_path_abs, bun.FD.cwd(), cache_path_abs).unwrap();
             } else {
                 defer file.close();
-                try bun.sys.renameat(tmpdir, tmp_path, cache_dir, outpath).unwrap();
+                try bun.sys.renameat(bun.toFD(tmpdir), tmp_path, bun.toFD(cache_dir), outpath).unwrap();
             }
         }